Fix workspace chat scroll jumps
