+++ packages/web-core/src/pages/kanban/ProjectRightSidebarContainer.tsx
@@ -25,6 +25,7 @@ import {
   ConversationList,
   type ConversationListHandle,
 } from '@/features/workspace-chat/ui/ConversationListContainer';
+import { usePinConversationToBottomOnChatBoxResize } from '@/features/workspace-chat/ui/usePinConversationToBottomOnChatBoxResize';
 import { RetryUiProvider } from '@/features/workspace-chat/model/contexts/RetryUiContext';
 import { createWorkspaceWithSession } from '@/shared/types/attempt';
 import { useAppNavigation } from '@/shared/hooks/useAppNavigation';
@@ -144,6 +145,7 @@ function WorkspaceSessionPanel({
   const routeState = useCurrentKanbanRouteState();
   const { workspaces: remoteWorkspaces } = useUserContext();
   const { activeWorkspaces, archivedWorkspaces } = useWorkspaceContext();
+  const containerRef = useRef<HTMLDivElement>(null);
   const conversationListRef = useRef<ConversationListHandle>(null);
   const [isAtBottom, setIsAtBottom] = useState(true);
   const { data: workspace, isLoading: isWorkspaceLoading } = useWorkspaceRecord(
@@ -230,6 +232,13 @@ function WorkspaceSessionPanel({
     setIsAtBottom(atBottom);
   }, []);
 
+  usePinConversationToBottomOnChatBoxResize({
+    containerRef,
+    conversationListRef,
+    isAtBottom,
+    scopeKey: `${workspaceId}:${selectedSessionId ?? 'new'}`,
+  });
+
   return (
     <ExecutionProcessesProvider
       key={`${workspaceId}-${selectedSessionId ?? 'new'}`}
@@ -238,7 +247,10 @@ function WorkspaceSessionPanel({
       <ApprovalFeedbackProvider>
         <EntriesProvider key={`${workspaceId}-${selectedSessionId ?? 'new'}`}>
           <MessageEditProvider>
-            <div className="relative flex h-full flex-1 flex-col bg-primary">
+            <div
+              ref={containerRef}
+              className="relative flex h-full flex-1 flex-col bg-primary"
+            >
               <div className="flex items-center justify-between px-base py-half border-b shrink-0">
                 <div className="flex items-center gap-half min-w-0 font-ibm-plex-mono">
                   <button
@@ -314,7 +326,10 @@ function WorkspaceSessionPanel({
                 </div>
               )}
 
-              <div className="flex justify-center @container pl-px">
+              <div
+                className="flex justify-center @container pl-px"
+                data-chatbox-container="true"
+              >
                 <SessionChatBoxContainer
                   {...(isSessionsLoading || isWorkspaceLoading
                     ? {
